Namespace all endpoints as /v1
Did this in the cli, the endpoint uris, and the rust modules.
Doing this any later in development will only get more annoying,
so now is the time.
