+++ agent/src/api/mod.rs
@@ -5,86 +5,10 @@ use axum::{
     response::Response,
 };
 use http::header::HeaderValue;
-use serde::{Deserialize, Serialize};
-use tokio::sync::Mutex;
-use utils::{blobs::BlobStore, errors::ServalError};
-use uuid::Uuid;
 
-use std::sync::Arc;
-use std::{collections::HashMap, path::PathBuf};
-
-pub mod jobs;
-pub mod proxy;
-pub mod storage;
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RunnerState {
-    instance_id: Uuid,
-    storage: Option<BlobStore>,
-    jobs: HashMap<String, JobMetadata>,
-    total: usize,
-    errors: usize,
-}
-
-impl RunnerState {
-    pub fn new(instance_id: Uuid, blob_path: Option<PathBuf>) -> Result<Self, ServalError> {
-        let storage = match blob_path {
-            Some(path) => Some(BlobStore::new(path)?),
-            None => None,
-        };
-
-        Ok(RunnerState {
-            instance_id,
-            storage,
-            total: 0,
-            errors: 0,
-            jobs: HashMap::new(),
-        })
-    }
-}
-
-pub type AppState = Arc<Mutex<RunnerState>>;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct JobMetadata {
-    id: Uuid,
-    name: String,
-    description: String,
-    status_url: String, // for the moment
-    result_url: String, // for the moment
-}
-
-impl JobMetadata {
-    pub fn new(name: String, description: String) -> Self {
-        let id = Uuid::new_v4();
-        Self {
-            id,
-            name,
-            description,
-            status_url: format!("/jobs/{}/status", id),
-            result_url: format!("/jobs/{}/result", id),
-        }
-    }
-}
-
-impl From<Envelope> for JobMetadata {
-    fn from(envelope: Envelope) -> Self {
-        let id = Uuid::new_v4();
-        Self {
-            id,
-            name: envelope.name.clone(),
-            description: envelope.description,
-            status_url: format!("/jobs/{}/status", id),
-            result_url: format!("/jobs/{}/result", id),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub struct Envelope {
-    name: String,
-    description: String,
-}
+pub mod v1;
+// Follow this pattern for additional major versions. E.g.,
+// pub mod v2;
 
 /// Remember what is important.
 pub async fn clacks<B>(req: Request<B>, next: Next<B>) -> Result<Response, StatusCode> {
