+++ agent/src/api/v1/mod.rs
@@ -1,0 +1,6 @@
+use http::header::HeaderValue;
+use uuid::Uuid;
+
+pub mod jobs;
+pub mod proxy;
+pub mod storage;
